+++ flask_accept/test/test_versioning.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from app_for_testing import app, index_without_fallback
+from flask_app_for_testing import app, index_without_fallback
 
 
 @pytest.mark.parametrize('headers,status_code,version', [
@@ -68,81 +68,3 @@ def test_with_double_wildcard(headers, status_code, rh):
         rv = c.get('/with-double-wildcard', headers={'accept': headers})
         if rv.status_code < 300:
             assert rh == json.loads(rv.data.decode())['rh']
-
-
-@pytest.mark.parametrize('headers,status_code,version', [
-    ('text/html', 200, 'v0'),
-    ('*/*', 200, 'v0'),
-    ('application/vnd.vendor.v1+json', 200, 'v1'),
-    ('application/json', 200, 'v2'),
-    ('application/json, text/html', 200, 'v2'),
-    ('application/vnd.vendor+json', 200, 'v2'),
-    ('application/vnd.vendor.v2+json', 200, 'v2'),
-])
-def test_flask_restful_resource_with_fallback(headers, status_code, version):
-    with app.test_client() as c:
-        rv = c.get('/resource/with-fallback', headers={'accept': headers})
-        assert rv.status_code == status_code
-        if rv.status_code < 300:
-            assert version == json.loads(rv.data.decode())['version']
-
-
-@pytest.mark.parametrize('headers,status_code,version', [
-    ('text/html', 406, None),
-    ('*/*', 406, None),
-    ('application/vnd.vendor.v1+json', 200, 'v1'),
-    ('application/json', 200, 'v2'),
-    ('application/json, text/html', 200, 'v2'),
-    ('application/vnd.vendor+json', 200, 'v2'),
-    ('application/vnd.vendor.v2+json', 200, 'v2'),
-])
-def test_flask_restful_resource_without_fallback(headers,
-                                                 status_code,
-                                                 version):
-    with app.test_client() as c:
-        rv = c.get('/resource/without-fallback', headers={'accept': headers})
-        assert rv.status_code == status_code
-        if rv.status_code < 300:
-            assert version == json.loads(rv.data.decode())['version']
-        else:
-            for accepted_type in index_without_fallback.accept_handlers:
-                assert accepted_type in rv.data.decode()
-
-
-@pytest.mark.parametrize('headers,status_code,version', [
-    ('text/html', 200, 'v0'),
-    ('*/*', 200, 'v0'),
-    ('application/vnd.vendor.v1+json', 200, 'v1'),
-    ('application/json', 200, 'v2'),
-    ('application/json, text/html', 200, 'v2'),
-    ('application/vnd.vendor+json', 200, 'v2'),
-    ('application/vnd.vendor.v2+json', 200, 'v2'),
-])
-def test_flask_restplus_resource_with_fallback(headers, status_code, version):
-    with app.test_client() as c:
-        rv = c.get('/plus/with-fallback', headers={'accept': headers})
-        assert rv.status_code == status_code
-        if rv.status_code < 300:
-            assert version == json.loads(rv.data.decode())['version']
-
-
-@pytest.mark.parametrize('headers,status_code,version', [
-    ('text/html', 406, None),
-    ('*/*', 406, None),
-    ('application/vnd.vendor.v1+json', 200, 'v1'),
-    ('application/json', 200, 'v2'),
-    ('application/json, text/html', 200, 'v2'),
-    ('application/vnd.vendor+json', 200, 'v2'),
-    ('application/vnd.vendor.v2+json', 200, 'v2'),
-])
-def test_flask_restplus_resource_without_fallback(headers,
-                                                  status_code,
-                                                  version):
-    with app.test_client() as c:
-        rv = c.get('/plus/without-fallback', headers={'accept': headers})
-        assert rv.status_code == status_code
-        if rv.status_code < 300:
-            assert version == json.loads(rv.data.decode())['version']
-        else:
-            for accepted_type in index_without_fallback.accept_handlers:
-                assert accepted_type in rv.data.decode()
